rename plugin config helper
