+++ lib/plugin_config.js
