+++ lib/run.js
@@ -1,11 +1,11 @@
-var idPlugin = require('./id_plugin');
+var pluginConfig = require('./plugin_config');
 
 module.exports = function run(name) {
   var action = require('../tasks/' + name);
   return function () {
     // call the action function with the plugin, then all
     // renaining arguments from commander
-    var plugin = idPlugin();
+    var plugin = pluginConfig();
     action.apply(null, [plugin].concat([].slice.apply(arguments)));
   };
 };
